Chore : sign_in/up renamed to auth.go & formatter executed
